Started to move module files into a new hierarchy.
